feat: restructure disclosure components
