+++ packages/components/src/components/disclosure/Accordion/types.ts
