+++ packages/components/src/components/disclosure/Tabs/index.ts
