+++ packages/components/src/components/disclosure/Tabs/types.ts
