+++ packages/components/src/components/disclosure/index.ts
@@ -1,0 +1,2 @@
+export * from './Accordion';
+export * from './Tabs';
