+++ packages/components/src/components/primitives/index.ts
@@ -1,7 +1,5 @@
-export * from './Accordion';
 export * from './Avatar';
 export * from './Card';
 export * from './CloseButton';
 export * from './Image';
 export * from './Icons';
-export * from './Tabs';
